TestInstanceTrackerFixture extracted
